+++ changelog.md
@@ -6,6 +6,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Added
+- Added collapsible help fields (See Issue #16 - @atmediauk)
+
 ## [0.4.0] - 2020-09-25
 
 ### Changed